// api/app/types/asaas_customer.ts
export interface CustomerProps {
  name: string
  cpfCnpj: string
  email: string
  phone?: string
  mobilePhone?: string
  address?: string
  addressNumber?: string
  complement?: string
  province?: string
  postalCode?: string
  externalReference?: string
  notificationDisabled?: boolean
  additionalEmails?: string
  municipalInscription?: string
  stateInscription?: string
  observations?: string
  groupName?: string | null
  company?: string | null
  foreignCustomer?: boolean
  userId?: number
}

export interface CustomerResponseCreataedProps {
  object: string
  id: string
  dateCreated: string
  name: string
  email: string
  phone: string
  mobilePhone: string
  address: string
  addressNumber: string
  complement: string
  province: string
  city: string
  cityName: string
  state: string
  country: string
  postalCode: string
  cpfCnpj: string
  personType: string
  deleted: boolean
  additionalEmails: string
  externalReference: string
  notificationDisabled: boolean
  observations: string
  foreignCustomer: boolean
}

export interface CustomerListProps {
  object: string
  hasMore: boolean
  totalCount: number
  limit: number
  offset: number
  data: CustomerResponseListProps[]
}

interface CustomerResponseListProps {
  object: string
  id: string
  dateCreated: string
  name: string
  email: string
  phone: string
  mobilePhone: string
  address: string
  addressNumber: string
  complement: string
  province: string
  city: string
  cityName: string
  state: string
  country: string
  postalCode: string
  cpfCnpj: string
  personType: string
  deleted: boolean
  additionalEmails: string
  externalReference: string
  notificationDisabled: boolean
  observations: string
  foreignCustomer: boolean
}